requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
dj-database-url==2.3.0
Django==5.2
django-autoslug==1.9.9
django-hosts==6.0
django-tailwind==4.0.1
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.3.2
propcache==0.3.1
psycopg==3.2.6
psycopg-binary==3.2.6
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.19.0
